Fix a number of compilation warnings

+++ benchmarks/matmul/matmul-lace.c
@@ -24,7 +24,7 @@ void init(REAL *A, int n)
 
     for (i = 0; i < n; i++) {
         for (j = 0; j < n; j++) {
-            A[i * n + j] = (double)rand();
+            A[i * n + j] = (REAL)rand();
         }
     }
 }
